Now compiles in Mac OS 10.5 or 10.6 SDK without warnings, other improvements and bug fixes

+++ NSString+VarArgs.m
@@ -23,7 +23,7 @@ static NSCharacterSet* static_modifierCharacterSet = nil ;
 		[static_modifierCharacterSet retain] ; // Stick around until app quits
 	}
 	
-	int possibleFormatCharLocation ;
+	NSUInteger possibleFormatCharLocation ;
 	NSString* modifier = @"" ;
 	BOOL dollarsign = [scanner scanString:@"$"intoString:NULL] ;
 	if (dollarsign) {
